Break out message types for Cryptoquote
